feat(ui): Add component to render Incident status changes
Status changes include: creating or detecting an incident, closing/reopening an incident

Fixes SEN-670
Fixes SEN-671
